Modified dataset loader
The dataset now accepts a sampling rate parameter. This parameter
is used to determine the interval between samples. It lets you pick
duration of observation/prediction. Longer observation periods are
sparse in position data, but they provide more context for the
model to learn prolonged motion patterns and vice versa.